Preliminary test cases passing.
Now we need new test cases to test for the `requireUpkeep` function. New test cases need to be formulated to test
`3-index approach`. Formal proofs need to be created and tested for. Finally, de-clutter the libraries and add a lot
more documentation as this is a neew kind of an approach.