[BUG FIX] `msgSender` replaced with `sender` from `agreementData`.
This bug is in relation to wrong units allocation and de-allocation in case an operator or external contract modifies a
stream. `msgSender` before this fix was that operator's or external contract's address due to which the shares of a user
whose stream is being modified were not deleted (in case of termination of stream). Furthermore, with an upcoming SF feature
(ACL), `msgSender` may not correspond to the actual owner of the stream. This bug fix will allow for compatibility with ACL
features. Further testing is required as this problem should have been caught in testing phase. The tests should check for
units allocation and de-allocation after every stream modification action.