[BUG-FIX] `calcUserUninvested` function reverted.
During testing, we found out that `calcUserUninvested` function which calculates uninvested supertoken amount of a user, reverted
in case a valid supertoken address wasn't provided to the function. Although this isn't a severe bug, it isn't expected from a
view function. I put in a simple condition which checks if the supertoken for which uninvested amount needs to be calculated is
initialised or not. If not inintialised, the function returns `0` as the answer. Also, the agreement hooks' helper functions now
expects supertoken address instead of underlying token address. This just makes more sense and avoids the above issue to a certain
extent.