Added delay parameter in buffer transfer amount calculation function.
Earlier commit didn't contain the `delay` parameter. A delay parameter is required to account for the delayed transaction times. As the buffer transfer amount depends upon the timing of the transaction execution, it's important to take enough allowance from the user in case buffer transfer amount needs to be taken from the user (upfront deposit) or else the transaction will fail.

Minor modifications to `dHedgeHelper.sol` library have also been made to remove some complexity and/or optimise gas usage.