New Feature Added (Subscription Index Prediction)
Everytime a user starts/updates/terminates a stream, he/she has to subscribe to an index for IDA distribution. However, the index to be subscribed to can only be calculated using data in the smart contract. Hence, it was easier to create a new function in the `dHedgeCore.sol` contract file. This will predict the index to be subscribed to in case of a stream action. New markets will be deployed with these changes.